+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Ability for Skills to write to Spaces (using External Nodes).
 - Add API support for creating and managing PaperQA collections.
+- Add support for different LLM providers by using litellm.
 
 ### Fixed